aaaa

+++ controllers/codeController.js
@@ -1,7 +1,7 @@
 const { findById } = require('../models/code');
 const Code = require('../models/code');
 const User = require('../models/user');
-
+const Order = require("../models/orders")
 exports.getAllCode = async (req, res) => {
     try {
         const code = await Code.find({});
@@ -22,6 +22,10 @@ exports.createCode = async (req, res) => {
         const { userID } = req.user;
         const admin = await User.findById(userID);
         if (admin.role === 'admin') {
+            const codeExisted = await Code.countDocuments({ code: req.body.code });
+            if (codeExisted > 0) {
+                return res.status(400).json({status: "failed", messenger: "Mã giảm giá đã tồn tại"});
+            }
             const code = await Code.create({
                 code: req.body.code,
                 discount: req.body.discount,
@@ -51,6 +55,14 @@ exports.deleteOneCode = async (req, res) => {
         const { userID } = req.user;
         const admin = await User.findById(userID);
         if (admin.role === 'admin') {
+            const count = await Order.countDocuments({ saleCode: req.params.id })
+
+            
+            if (count > 0) {
+                return res.status(400).json({
+                    messenger: "Voucher đã được đặt nên không xóa được"
+                });
+            }
             await Code.findByIdAndDelete(req.params.id);
             res.json({
                 status: "success"
@@ -87,6 +99,18 @@ exports.getOneCode = async (req, res) => {
 exports.editOneCode = async (req, res) => {
     try {
         const { id } = req.params;
+        const { code } = req.body;
+
+        // Kiểm tra xem mã giảm giá mới có trùng với mã giảm giá đã tồn tại hay không
+        const existingCode = await Code.findOne({ code: code, _id: { $ne: id } });
+        if (existingCode) {
+            return res.status(400).json({
+                status: 'failed',
+                message: 'Mã giảm giá đã tồn tại'
+            });
+        }
+
+        // Cập nhật mã giảm giá
         await Code.findByIdAndUpdate(id, { ...req.body }, { runValidators: true, new: true });
         res.json({
             status: "success"
